requirements.txt
tqdm
opencv-python
python-dotenv
torch
torchvision
pillow
numpy
pandas
scikit-learn
matplotlib
seaborn
pytest
mlflow
fastapi
uvicorn
python-multipart
pyyaml
hydra-core